requirements.txt
flask
gunicorn
python-binance
pyTelegramBotAPI
python-telegram-bot
